先传一下，有一些TODO

+++ LogisticManageSystem/Client/src/main/java/edu/nju/lms/businessLogicService/impl/department/DepartmentblImpl.java
@@ -26,6 +26,7 @@ public class DepartmentblImpl{
 		try {
 			departmentPO = service.findDepartment(id);
 		} catch (RemoteException e) {
+			//TODO
 		}
 		if(departmentPO!=null) {
 			result=new DepartmentVO(departmentPO.getType(), departmentPO.getDepartmentNum(), departmentPO.getLocation());
@@ -34,16 +35,25 @@ public class DepartmentblImpl{
 	}
 
 	public ResultMessage deleteDepartment(String id) {
-		ResultMessage message = new ResultMessage(false, "网络未连接");
+		ResultMessage message = idCheck(id);
+		if(!message.isSuccess()){
+			return message;
+		}
+		message = new ResultMessage(false, "网络未连接");
 		try {
 			message=service.deleteDepartment(id);
 		} catch (RemoteException e) {
+			//TODO
 		}
 		return message;
 	}
 
 	public ResultMessage updateDepartment(DepartmentVO department) {
-		ResultMessage message = new ResultMessage(false, "网络未连接");
+		ResultMessage message = idCheck(department.getDepartmentNum());
+		if(!message.isSuccess()){
+			return message;
+		}
+		message = new ResultMessage(false, "网络未连接");
 		CityPO city = null;
 		DepartmentPO departmentPO = null;
 		try {
@@ -55,12 +65,17 @@ public class DepartmentblImpl{
 				message = service.updateDepartment(departmentPO);
 			}
 		} catch (RemoteException e) {
+			//TODO
 		}
 		return message;
 	}
 
 	public ResultMessage addDepartment(DepartmentVO department) {
-		ResultMessage message = new ResultMessage(false, "网络未连接");
+		ResultMessage result=idCheck(department.getDepartmentNum());
+		if(!result.isSuccess()){
+			return result;
+		}
+		result = new ResultMessage(false, "网络未连接");
 		CityPO city = null;
 		DepartmentPO departmentPO = null;
 		try {
@@ -70,19 +85,25 @@ public class DepartmentblImpl{
 			} else {
 				departmentPO = new DepartmentPO(department.getType(), 
 						department.getDepartmentNum(), city.getId());
-				message = service.addDepartment(departmentPO);
+				result = service.addDepartment(departmentPO);
 			}
 		} catch (RemoteException e) {
+			//TODO
 		}
-		return message;
+		return result;
 	}
 
 	public ResultMessage addCity(CityVO city) {
-		ResultMessage result=new ResultMessage(false,"网络未连接");
+		ResultMessage result=idCheck(city.getId());
+		if(!result.isSuccess()){
+			return result;
+		}
 		CityPO cityPO=new CityPO(city.getId(),city.getName(),city.getBusinessNums(),city.getDistance());
+		result=new ResultMessage(false,"网络未连接");
 		try {
 			result=service.addCity(cityPO);
 		} catch (RemoteException e) {
+			//TODO
 		}
 		return result;
 	}
@@ -93,6 +114,7 @@ public class DepartmentblImpl{
 		try {
 			cityPO=service.findCity(id);
 		} catch (RemoteException e) {
+			//TODO
 		}
 		if(cityPO.getId()!=null){
 			city=new CityVO(cityPO.getId(),cityPO.getName(),cityPO.getBusinessNums(),cityPO.getDistance());
@@ -100,4 +122,13 @@ public class DepartmentblImpl{
 		return city;
 	}
 
+	public ResultMessage idCheck(String id){
+		ResultMessage result=new ResultMessage(true,"");
+		if(id.length()!=6){
+			result.setSuccess(false);
+			result.setErrorMessage("输入机构的位数不正确！");
+			return result;
+		}
+		return result;
+	}
 }
